Changed some code during live refactoring
